refactor to cleanup
