+++ src/FormLayout/Generic.php
@@ -9,13 +9,8 @@ use atk4\ui\View;
 /**
  * Generic Layout for a form.
  */
-class Generic extends View
+class Generic extends _Abstract
 {
-    /**
-     * Links layout to the form.
-     */
-    public $form = null;
-
     // @var inheritdoc
     public $defaultTemplate = 'formlayout/generic.html';
 
@@ -35,118 +30,8 @@ class Generic extends View
      */
     public $inline = null;
 
-    /**
-     * Places field inside a layout somewhere. Should be called
-     * through $form->addField().
-     *
-     * @param string|null              $name
-     * @param array|string|object|null $decorator
-     * @param array|string|object|null $field
-     *
-     * @return \atk4\ui\FormField\Generic
-     */
-    public function addField($name, $decorator = null, $field = null)
-    {
-        if (!is_string($name)) {
-            throw new Exception(['Format for addField now require first argument to be name']);
-        }
-
-        if (!$this->form->model) {
-            $this->form->model = new \atk4\ui\misc\ProxyModel();
-        }
-
-        if (is_string($field)) {
-            $field = ['type' => $field];
-        }
-
-        if ($name) {
-            $existingField = $this->form->model->hasElement($name);
-        }
-
-        try {
-            if (!$existingField) {
-                // Add missing field
-                if ($field) {
-                    $field = $this->form->model->addField($name, $field);
-                } else {
-                    $field = $this->form->model->addField($name);
-                }
-            } elseif (is_array($field)) {
-                // Add properties to existing field
-                $existingField->setDefaults($field);
-                $field = $existingField;
-            } elseif (is_object($field)) {
-                throw new Exception(['Duplicate field', 'name' => $name]);
-            } else {
-                $field = $existingField;
-            }
-
-            if (is_string($decorator)) {
-                $decorator = $this->form->decoratorFactory($field, ['caption' => $decorator]);
-            } elseif (is_array($decorator)) {
-                $decorator = $this->form->decoratorFactory($field, $decorator);
-            } elseif (!$decorator) {
-                $decorator = $this->form->decoratorFactory($field);
-            } elseif (is_object($decorator)) {
-                if (!$decorator instanceof \atk4\ui\FormField\Generic) {
-                    throw new Exception(['Field decorator must descend from \atk4\ui\FormField\Generic', 'decorator' => $decorator]);
-                }
-                $decorator->field = $field;
-                $decorator->form = $this->form;
-            } else {
-                throw new Exception(['Value of $decorator argument is incorrect', 'decorator' => $decorator]);
-            }
-        } catch (\Throwable $e) {
-            throw new Exception(['Unable to add form field', 'name' => $name, 'decorator' => $decorator, 'field' => $field], null, $e);
-        }
-
+    protected function _addField($decorator, $field) {
         return $this->_add($decorator, ['desired_name' => $field->short_name]);
-    }
-
-    public function setModel(\atk4\data\Model $model, $fields = null)
-    {
-        parent::setModel($model);
-
-        if ($fields === false) {
-            return $model;
-        }
-
-        if ($fields === null) {
-            $fields = [];
-            foreach ($model->elements as $f) {
-                if (!$f instanceof \atk4\data\Field) {
-                    continue;
-                }
-
-                if (!$f->isEditable()) {
-                    continue;
-                }
-                $fields[] = $f->short_name;
-            }
-        }
-
-        if (is_array($fields)) {
-            foreach ($fields as $field) {
-                $this->addField($field);
-            }
-        } else {
-            throw new Exception(['Incorrect value for $fields', 'fields' => $fields]);
-        }
-
-        return $model;
-    }
-
-    /**
-     * Return Field decorator associated with
-     * the form's field.
-     */
-    public function getField($name)
-    {
-        if (empty($this->form)) {
-            throw new Exception(['Incorrect value for $form', 'form' => $this->form]);
-        }
-
-        return $this->form->getField($name);
     }
 
     /**
